+++ lib/models/test_action.dart
@@ -1,128 +1,140 @@
-import 'dart:async';
-
+import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:test_actions/models/test_action_type.dart';
 
-///Test action, class that can be performed on a [WidgetTester].
+///Class to perform a certain [ActionType] on the [WidgetTester]
 class TestAction {
-  final TestActionType actionType;
+  ///Type of the action to be performed on the [Finder] widget.
+  final TestActionType action;
 
-  ///Type [Duration].
+  ///Only needed if the tester is not set in the [TestActions] class or is executed outside the cited class.
   ///
-  ///Can be used as [Duration] for the pump.
-  ///
-  ///Throws an error if the [TestActionType] is AwaitFuture, but this is null.
-  final Duration? awaitDuration;
+  ///Fails the test if this condition is not satisfied.
+  final WidgetTester? tester;
 
-  ///Type [int]
+  ///Needed when [TestActionType] != PumpAndSettle.
   ///
-  ///How many times the pump should be called
-  ///
-  ///The default value is 30
-  final int pumpTimes;
-
-  ///The finder used to find the widget to be pressed
+  ///Fails the test if this condition is not satisfied.
   final Finder? finder;
 
-  ///Custom function to be called
+  ///Needed when [TestActionType] == EnterText.
+  ///
+  ///Fails the test if this condition is not satisfied.
+  final String? enterText;
+
+  ///Needed when [TestActionType] == AwaitFuture
+  ///
+  ///Can be used as the [Duration] of the Pump Action
+  ///
+  ///Fails the test if this condition is not satisfied.
+  final Duration? awaitDuration;
+
+  ///Needed when [TestActionType] == CustomAction
+  ///
+  ///Fails the test if this condition is not satisfied.
   final Function? customAction;
 
-  ///The [WidgetTester] to performs the actions on.
-  ///
-  ///Should never be null
-  late WidgetTester? tester;
-
-  ///Decides if the pumpAndSettle should be run at the end of the action.
-  ///
-  ///The default value is true.
+  ///Defaults to true, controls the execution of the [WidgetTester].pumpAndSettle at the end of the action.
   final bool executePumpAndSettle;
 
+  ///How many times the pump should be called.
+  ///
+  /// Default to 5
+  final int pumpTime;
+
+  /// The [Offset] used to drag a [Widget] using the [Finder]
+  ///
+  /// Needed when the [TestActionType] == Drag
+  final Offset? dragOffset;
+
   TestAction({
-    required this.actionType,
-    this.awaitDuration,
-    this.pumpTimes = 30,
-    this.finder,
+    required this.action,
     this.tester,
+    this.finder,
+    this.enterText,
+    this.awaitDuration,
     this.customAction,
     this.executePumpAndSettle = true,
+    this.pumpTime = 5,
+    this.dragOffset,
   });
 
-  FutureOr<dynamic> performAction(
-      int actionIndex, Function setActionAsDone) async {
-    print('Performing ${actionType.toString().split('.')[1]} #$actionIndex');
+  // : assert(
+  //     action == ActionType.EnterText
+  //         ? enterText != null
+  //         : enterText == null &&
+  //                 ActionType.values.contains(action) &&
+  //                 action == ActionType.PumpAndSettle
+  //             ? finder == null
+  //             : finder != null && action == ActionType.TestExpect
+  //                 ? neededValue != null && valueToCheck != null
+  //                 : neededValue == null &&
+  //                         valueToCheck == null &&
+  //                         action == ActionType.FutureAwait
+  //                     ? awaitDuration != null
+  //                     : awaitDuration == null &&
+  //                             action == ActionType.CustomAction
+  //                         ? customAction != null
+  //                         : customAction == null,
+  //   );
+
+  ///Performs the action
+  Future<void> performAction(
+      {required Function setAsDone, required int actionIndex}) async {
+    print('Performing ${action.toString().split('.')[1]} action #$actionIndex');
     try {
-      switch (actionType) {
-        case TestActionType.PumpAndSettle:
-          //Executes the pumpAndSettle
-          await tester!.pumpAndSettle();
-          break;
-        case TestActionType.Pump:
-          bool pumpDone = false;
-          for (int i = 0; i < pumpTimes; i++) {
-            await tester!.pump(awaitDuration);
-            if (i == pumpTimes - 1) pumpDone = true;
-          }
-          //Awaits all the pumps to be done
-          while (!pumpDone) {
-            await Future.delayed(awaitDuration ?? Duration(milliseconds: 100));
-          }
-          break;
-        case TestActionType.CustomAction:
-          //Executes the custom action
-          customAction!();
-          break;
+      switch (action) {
         case TestActionType.Press:
-          //Executes the tap action
           await tester!.tap(finder!);
           break;
-        case TestActionType.AwaitFuture:
-          //Awaits for a certain Duration
+        case TestActionType.EnterText:
+          await tester!.enterText(finder!, enterText!);
+          break;
+        case TestActionType.PumpAndSettle:
+          await tester!.pumpAndSettle();
+          break;
+        case TestActionType.FutureAwait:
+          print(
+              '** Future delayed by ${awaitDuration!.inSeconds}s -> ${awaitDuration!.inMilliseconds}ms');
           await Future.delayed(awaitDuration!);
           break;
-        default:
-          throw Exception('This action has not been yet implemented');
+        case TestActionType.CustomAction:
+          await customAction!();
+          break;
+        case TestActionType.Pump:
+          print('* Pumping for $pumpTime times');
+          bool pumpDone = false;
+          for (int i = 0; i < pumpTime; i++) {
+            await tester!.pump(Duration(milliseconds: 100));
+            if (i == pumpTime - 1) pumpDone = true;
+          }
+          do {
+            await Future.delayed(Duration(milliseconds: 100));
+          } while (!pumpDone);
+          setAsDone(actionIndex);
+          print('Action ${action.toString().split('.')[1]} #$actionIndex done');
+          return;
+        case TestActionType.Drag:
+          await tester!.drag(finder!, dragOffset!);
+          break;
       }
+      if (executePumpAndSettle && action != TestActionType.EnterText) {
+        print('Executing pumpAndSettle');
+        await tester!.pumpAndSettle(Duration(milliseconds: 500));
+      } else
+        print('* Avoiding the pumpAndSettle');
     } catch (e) {
-      _printError(e, actionIndex, surplusMessage: 'While running the action');
+      printError(e, actionIndex);
       return;
     }
-    //Tries to run the pumpAndSettle if {executePumpAndSettle} is true
-    try {
-      if (executePumpAndSettle) {
-        print('=> executing pumpAndSettle');
-        await tester!.pumpAndSettle();
-      }
-    } catch (e) {
-      _printError(e, actionIndex,
-          surplusMessage: 'While running the tester.pumpAndSettle()');
-    }
-    print(
-        'Action ${actionType.toString().split('.')[1]} #$actionIndex performed');
-    setActionAsDone(actionIndex);
+    setAsDone(actionIndex);
+    print('Action ${action.toString().split('.')[1]} #$actionIndex done');
+    return;
   }
 
-  TestAction copyWith({
-    TestActionType? actionType,
-    WidgetTester? tester,
-    int? pumpTimes,
-    Duration? awaitDuration,
-    Finder? finder,
-    Function? customAction,
-    bool? executePumpAndSettle,
-  }) =>
-      TestAction(
-        actionType: actionType ?? this.actionType,
-        tester: tester ?? this.tester,
-        pumpTimes: pumpTimes ?? this.pumpTimes,
-        awaitDuration: awaitDuration ?? this.awaitDuration,
-        finder: finder ?? this.finder,
-        customAction: customAction ?? this.customAction,
-        executePumpAndSettle: executePumpAndSettle ?? this.executePumpAndSettle,
-      );
-
-  void _printError(dynamic e, int actionIndex, {String surplusMessage = ''}) {
+  void printError(dynamic e, int actionIndex, {String surplusMessage = ''}) {
     String firstMessage =
-        '* Error accoured while performing ${actionType.toString().split('.')[1]} action #$actionIndex';
+        '* Error accoured while performing ${action.toString().split('.')[1]} action #$actionIndex';
     String secondMessage =
         '* ErrorType: ${e.runtimeType}\t Message: ${e.toString()}';
     String divider = '';
@@ -138,4 +150,27 @@ class TestAction {
     if (surplusMessage != '') print('* $surplusMessage');
     print(divider);
   }
+
+  TestAction copyWith({
+    TestActionType? actionType,
+    WidgetTester? tester,
+    Finder? finder,
+    String? enterText,
+    Duration? awaitFuture,
+    Function? customAction,
+    bool? executePumpAndSettle,
+    Offset? dragOffset,
+    int? pumpTime,
+  }) =>
+      TestAction(
+        action: actionType ?? this.action,
+        executePumpAndSettle: executePumpAndSettle ?? this.executePumpAndSettle,
+        awaitDuration: awaitDuration ?? this.awaitDuration,
+        finder: finder ?? this.finder,
+        enterText: enterText ?? this.enterText,
+        tester: tester ?? this.tester,
+        customAction: customAction ?? this.customAction,
+        dragOffset: dragOffset ?? this.dragOffset,
+        pumpTime: pumpTime ?? this.pumpTime,
+      );
 }