+++ test/test_actions_test.dart
@@ -6,17 +6,17 @@ import 'package:mockito/mockito.dart';
 class WidgetTesterMock extends Mock implements WidgetTester {}
 
 void main() async {
-  WidgetTesterMock mockTester = WidgetTesterMock();
+  WidgetTesterMock mockTester;
   late TestActions actions;
 
   setUp(() {
-    actions = TestActions();
-    actions.tester = mockTester;
+    mockTester = WidgetTesterMock();
+    actions = TestActions(tester: mockTester);
     actions.resetActions();
   });
 
   group('TestAction test', () {
-    test('out of bound action perform', () async {
+    test('[out of bound action perform]', () async {
       try {
         await actions.performActionAt(2);
       } catch (e) {
@@ -24,30 +24,30 @@ void main() async {
       }
     });
 
-    test('perform all actions when list is empty', () async {
-      dynamic result = await actions.performAllActions();
-      expect(result, isNull);
+    test('[perform all actions when list is empty]', () async {
+      await actions.performActions();
+      expect(actions.areAllDone(), true);
     });
 
-    test('adding one action and then reset', () {
-      actions.addAction(TestAction(actionType: TestActionType.AwaitFuture));
+    test('[adding one action and then reset]', () {
+      actions.addAction(TestAction(action: TestActionType.FutureAwait));
       expect(actions.actions.length, 1);
       actions.resetActions();
       expect(actions.actions.length, 0);
       expect(actions.tester, isNotNull);
     });
-    test('adding more actions', () {
+    test('[adding more actions]', () {
       addSampleActions(actions);
       expect(actions.actions.length, 4);
     });
-    test('running a single custom action', () async {
+    test('[running a single custom action]', () async {
       addSampleActions(actions);
       await actions.performActionAt(1);
       expect(actions.isActionDone(1), true);
     });
-    test('run all the actions', () async {
+    test('[run all the actions]', () async {
       addSampleActions(actions);
-      await actions.performAllActions();
+      await actions.performActions();
       for (int i = 0; i < actions.actions.length; i++) {
         expect(actions.isActionDone(i), true);
       }
@@ -58,35 +58,35 @@ void main() async {
 void addSampleActions(TestActions actions, {bool addOne = false}) {
   if (addOne)
     actions.addAction(TestAction(
-      actionType: TestActionType.CustomAction,
+      action: TestActionType.CustomAction,
       executePumpAndSettle: false,
       customAction: () {
         print('This is my test custom action');
       },
     ));
   else
-    actions.addMultipleActions([
+    actions.addActionsAll([
       TestAction(
-        actionType: TestActionType.AwaitFuture,
+        action: TestActionType.FutureAwait,
         executePumpAndSettle: false,
         awaitDuration: Duration(seconds: 2),
       ),
       TestAction(
-        actionType: TestActionType.CustomAction,
+        action: TestActionType.CustomAction,
         executePumpAndSettle: false,
         customAction: () {
           print('This is my test custom action');
         },
       ),
       TestAction(
-        actionType: TestActionType.CustomAction,
+        action: TestActionType.CustomAction,
         executePumpAndSettle: false,
         customAction: () {
           print('This is my second test for a custom action');
         },
       ),
       TestAction(
-        actionType: TestActionType.AwaitFuture,
+        action: TestActionType.FutureAwait,
         executePumpAndSettle: false,
         awaitDuration: Duration(seconds: 2),
       ),